feat(build): update includes path for source files and add compile flags to file

+++ src/util/str.h
@@ -4,6 +4,6 @@
 #include <stddef.h>
 
 void *nf_memcpy(void *dst, const void *src, size_t n);
-int getlen(const char *str);
+size_t getlen(const char *str);
 
 #endif
